test(vm-driver): contract checks for prepare, prepared listeners, discard and disk reattachment
